All working! Using firebase realtime database for the notifications about the alerts, as well as sending the alerts.
Now it's possible to generate new simples tests alerts in the OptionScreen, so now we know how to insert new alerts in the database.

It's also possible to go back from the OptionScreen to the HomeScreen, and from the NewUserFormScreen to the OptionScreen.

The alerts the app receives can be received with the app in the background.

new files created:
Alert.kt
AlertService.kt
Changed a little bit the AlertFireB.kt

TODO
- Work on the alerts cards, and start creating a custom notification message
- Understand how to make the app receive notifications when the app is closed.